Implement back-end ability to set priority
There's been an issue with inconsistency in the database names that's been giving me a lot of grief for the past two or so class periods... it's incredibly important that we remember to make sure we're reading from the correct databases in our backend code. There is a LOT of conflict between the work we're pulling from both of the previous repositories for iteration 2. I still need to sort some things out before this is fully ready for use.